Fix empty hostname bug
The read statement will evaulate the lines[n++] expression for
each line in the input wordlist that does not start with #. It
will however also evaluate the expression when it reads an EOF
from the stream. This means that n will equal the number of
entries in lines plus one, which in turn means that when rand
returns n-1 in r, the resulting line will be empty.

Signed-off-by: Jonas Eriksson <zqad@acc.umu.se>