Avoid the same hostname in one session
By adding a # in front of used words, genhost will not display
the same word in two subsequent runs. It is however possible to
display the same word within one run. When generating a large
number of hosts, it might be hard to notice the duplication.
This commit solves this issue by filtering the wordlist in each
iteration. The code for this is not very beautiful, since bash
arrays are essentially associative arrays.

Signed-off-by: Jonas Eriksson <zqad@acc.umu.se>